Add OrderTable, ProfileImage, and ProfileDetails components; refactor Profile view to integrate new components and enhance user experience
